Extract shared fleet record predicates to break the count-collapse import cycle.
Centralize ship-id matching, direct-observation detection, and known built-turn lookup so observation ingest can import collapse without a lazy cycle or duplicated helpers.
